Move directory_contents to utils
